feat: Refactor due amount handling in ApplicationsTable
- Extracted the due amount change logic from the ApplicationsTable component into a new custom hook, useDueAmountChange, for improved separation of concerns and code organization.
- The new hook includes validation and state management for updating the due amount, enhancing maintainability and clarity in the ApplicationsTable component.
